Rewrite bela::process::Process LookupPath
