+++ include/bela/process.hpp
@@ -2,8 +2,8 @@
 #ifndef BELA_PROCESS_HPP
 #define BELA_PROCESS_HPP
 #include "base.hpp"
-#include "env.hpp"
 #include "escapeargv.hpp"
+#include "simulator.hpp"
 
 namespace bela::process {
 constexpr const wchar_t *string_nullable(std::wstring_view str) {
@@ -17,22 +17,14 @@ enum CaptureMode {
   CAPTURE_USEIN = 0x2,
   CAPTURE_USEERR = 0x4,
 };
-
+using bela::env::Simulator;
 class Process {
 public:
-  Process() = default;
+  Process(const Simulator *simulator_ = nullptr) : simulator(simulator_) {}
   Process(const Process &) = delete;
   Process &operator=(const Process &) = delete;
   Process &Chdir(std::wstring_view dir) {
     cwd = dir;
-    return *this;
-  }
-  Process &SetEnv(std::wstring_view key, std::wstring_view val, bool force = false) {
-    de.SetEnv(key, val, force);
-    return *this;
-  }
-  Process &SetEnvStrings(std::wstring_view env_) {
-    env = env_;
     return *this;
   }
   const bela::error_code &ErrorCode() const { return ec; }
@@ -40,28 +32,27 @@ public:
   std::string_view Out() const { return out; } // when call Capture
   template <typename... Args> int Execute(std::wstring_view cmd, Args... args) {
     bela::EscapeArgv ea(cmd, args...);
-    exitcode = ExecuteInternal(ea.data());
+    exitcode = ExecuteInternal(cmd, ea.data());
     return exitcode;
   }
   template <typename... Args> int Capture(std::wstring_view cmd, Args... args) {
     bela::EscapeArgv ea(cmd, args...);
-    exitcode = CaptureInternal(ea.data(), CAPTURE_OUT);
+    exitcode = CaptureInternal(cmd, ea.data(), CAPTURE_OUT);
     return exitcode;
   }
   template <typename... Args>
   int CaptureWithMode(DWORD flags, std::wstring_view cmd, Args... args) {
     bela::EscapeArgv ea(cmd, args...);
-    exitcode = CaptureInternal(ea.data(), flags);
+    exitcode = CaptureInternal(cmd, ea.data(), flags);
     return exitcode;
   }
 
 private:
-  int ExecuteInternal(wchar_t *cmdline);
-  int CaptureInternal(wchar_t *cmdline, DWORD flags);
+  int ExecuteInternal(std::wstring_view file, wchar_t *cmdline);
+  int CaptureInternal(std::wstring_view file, wchar_t *cmdline, DWORD flags);
+  const Simulator *simulator{nullptr};
   std::wstring cwd;
-  std::wstring env;
   std::string out;
-  bela::env::Derivator de;
   bela::error_code ec;
   DWORD pid{0};
   DWORD exitcode{0};
