+++ test/winutils/envps.cc
@@ -1,4 +1,4 @@
-#include <bela/env.hpp>
+#include <bela/simulator.hpp>
 
 int LinkToApp(wchar_t *env) {
   STARTUPINFOW si;
@@ -22,9 +22,10 @@ int LinkToApp(wchar_t *env) {
 }
 
 int wmain() {
-  bela::env::Derivator dev;
-  dev.SetEnv(L"GOPROXY", L"https://goproxy.io/");
-  dev.SetEnv(L"Path", L"C:/Dev");
-  auto envs = dev.CleanupEnv(L"C:/Go");
+  bela::env::Simulator simulator;
+  simulator.InitializeCleanupEnv();
+  simulator.SetEnv(L"GOPROXY", L"https://goproxy.io/");
+  // simulator.SetEnv(L"Path", L"C:/Dev");
+  auto envs = simulator.MakeEnv();
   return LinkToApp(envs.data());
 }
